+++ bearhub/__init__.py
@@ -1,6 +1,6 @@
 import os
 
-__version__ = '0.10.7'
+__version__ = '0.10.8'
 __app_name__ = 'bearhub'
 
 ROOT_DIR = os.path.dirname(os.path.abspath(__file__))